+++ src/main/webapp/components/histogram/index.js
@@ -1,0 +1,2 @@
+import Histogram from './histogram'
+export { Histogram as default }
